+++ packages/db/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 0.36.0 - 2024-10-07
+
+### Changed
+
+- NPostgres `parseEventRow` method is protected.
+
 ## 0.35.0 - 2024-09-23
 
 ### Added